requirements.txt
gunicorn==20.1.0
asgiref==3.7.2
distlib==0.3.6
Django==4.2.2
django-seed==0.3.1
django-urls==1.1.3
djangorestframework==3.14.0
Faker==18.10.1
filelock==3.12.0
pipreq==0.4
platformdirs==3.5.1
PyMySQL==1.0.3
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
toposort==1.10
virtualenv==20.23.0
